nuevo

+++ src/main/java/com/fisisac/developer/fisisackback/domain/Programa.java
@@ -14,7 +14,7 @@ public class Programa {
   @Id
   @Column(name = "id_programa")
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Long id;
+  private Short id;
   @Column(name = "nom_programa")
   private String nombrePrograma;
   @Column(name = "sigla_programa")
@@ -27,7 +27,7 @@ public class Programa {
   public Programa() {
   }
 
-  public Programa(Long id, String nombrePrograma, String siglaPrograma, Boolean vigenciaPrograma,
+  public Programa(Short id, String nombrePrograma, String siglaPrograma, Boolean vigenciaPrograma,
       Integer numeroPrioridad) {
     this.id = id;
     this.nombrePrograma = nombrePrograma;
@@ -36,11 +36,11 @@ public class Programa {
     this.numeroPrioridad = numeroPrioridad;
     }
 
-  public Long getId() {
+  public Short getId() {
     return this.id;
   }
 
-  public void setId(Long id) {
+  public void setId(Short id) {
     this.id = id;
   }
